Adaptation according by github. Refactored code into the helper method (FCM/FCMToken.php -> BuildFCM).
