+++ app/Http/Controllers/FCMTokenController.php
@@ -5,14 +5,16 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\User;
 use App\Token;
+use Illuminate\Support\Facades\Auth;
 
 class FCMTokenController extends Controller
 {
     public function addToken(Request $request)
     {
+        $user = Auth::user();
         if ($request->has('FCMToken')) {
             $FCMToken = new Token();
-            $FCMToken->user_id = User::getUserIdFromToken($request);
+            $FCMToken->user_id = $user->id;
             $FCMToken->token = $request->FCMToken;
             $FCMToken->save();
             return response()->json(["proved" => true, "message" => "Token has been added"], 200);
